refactor(preact-template): upgrading preact to use preact x
