+++ packages/teleport-plugin-common/src/utils/ast-utils.ts
@@ -1,6 +1,7 @@
 import * as types from '@babel/types'
 import ParsedASTNode from './parsed-ast'
 import { StringUtils } from '@teleporthq/teleport-shared'
+import { UIDLStateDefinition, UIDLPropDefinition } from '@teleporthq/teleport-types'
 /**
  * Adds a class definition string to an existing string of classes
  */
@@ -290,4 +291,114 @@ export const removeAttributeByName = (jsxTag: types.JSXElement, attrName: string
       attr.type === 'JSXSpreadAttribute' ||
       (attr.type === 'JSXAttribute' && attr.name.name !== attrName)
   )
+}
+
+export const createClassComponent = (
+  name: string,
+  propDefinitions: Record<string, UIDLPropDefinition>,
+  stateDefinitions: Record<string, UIDLStateDefinition>,
+  jsxTagTree: types.JSXElement,
+  t = types
+) => {
+  // TODO: Add event handlers as separate functions later
+  const classMethodsAndProperties = []
+  const renderMethodArguments = []
+
+  if (Object.keys(propDefinitions).length > 0 || Object.keys(stateDefinitions).length > 0) {
+    renderMethodArguments.push(t.identifier('props'))
+  }
+
+  if (Object.keys(stateDefinitions).length > 0) {
+    const stateDeclarationsAST = Object.keys(stateDefinitions).map((stateKey) => {
+      const stateDefinition = stateDefinitions[stateKey]
+      return t.objectProperty(
+        t.identifier(stateKey),
+        convertValueToLiteral(stateDefinition.defaultValue)
+      )
+    })
+
+    classMethodsAndProperties.push(
+      t.classProperty(t.identifier('state'), t.objectExpression(stateDeclarationsAST))
+    )
+    renderMethodArguments.push(t.identifier('state'))
+  }
+
+  const classBody = t.classBody([
+    ...classMethodsAndProperties,
+    t.classMethod(
+      'method',
+      t.identifier('render'),
+      renderMethodArguments,
+      t.blockStatement([t.returnStatement(jsxTagTree)])
+    ),
+  ])
+
+  const classDeclaration = t.classDeclaration(
+    t.identifier(name),
+    t.identifier('Component'),
+    classBody,
+    null
+  )
+
+  return classDeclaration
+}
+
+export const createPureComponent = (
+  name: string,
+  stateDefinitions: Record<string, UIDLStateDefinition>,
+  propDefinitions: Record<string, UIDLPropDefinition>,
+  jsxTagTree: types.JSXElement,
+  t = types
+): types.VariableDeclaration => {
+  const arrowFunctionBody = createReturnExpressionSyntax(stateDefinitions, jsxTagTree)
+  const arrowFunctionProps = []
+
+  if (Object.keys(propDefinitions).length > 0) {
+    arrowFunctionProps.push(t.identifier('props'))
+  }
+
+  const arrowFunction = t.arrowFunctionExpression(arrowFunctionProps, arrowFunctionBody)
+
+  const declarator = t.variableDeclarator(t.identifier(name), arrowFunction)
+  const component = t.variableDeclaration('const', [declarator])
+
+  return component
+}
+
+export const createReturnExpressionSyntax = (
+  stateDefinitions: Record<string, UIDLStateDefinition>,
+  jsxTagTree: types.JSXElement,
+  t = types
+) => {
+  const returnStatement = t.returnStatement(jsxTagTree)
+
+  const stateHooks = Object.keys(stateDefinitions).map((stateKey) =>
+    createStateHookAST(stateKey, stateDefinitions[stateKey])
+  )
+
+  return t.blockStatement([...stateHooks, returnStatement] || [])
+}
+
+/**
+ * Creates an AST line for defining a single state hook
+ */
+export const createStateHookAST = (
+  stateKey: string,
+  stateDefinition: UIDLStateDefinition,
+  t = types
+) => {
+  const defaultValueArgument = convertValueToLiteral(
+    stateDefinition.defaultValue,
+    stateDefinition.type
+  )
+
+  return t.variableDeclaration('const', [
+    t.variableDeclarator(
+      t.arrayPattern([
+        t.identifier(stateKey),
+        t.identifier(`set${StringUtils.capitalize(stateKey)}`),
+      ]),
+      t.callExpression(t.identifier('useState'), [defaultValueArgument])
+    ),
+  ])
 }
