+++ packages/teleport-plugin-preact-base-component/src/constants.ts
@@ -1,9 +1,18 @@
 import { UIDLDependency } from '@teleporthq/teleport-types'
 
-export const PREACT_COMPONENT_DEPENDENCY: UIDLDependency = {
+export const PREACTJSX_PRAGMA_DEPENDENCY: UIDLDependency = {
   type: 'library',
   path: 'preact',
   version: 'next',
+  meta: {
+    namedImport: true,
+  },
+}
+
+export const USE_STATE_DEPENDENCY: UIDLDependency = {
+  type: 'library',
+  path: 'preact/hooks',
+  version: '16.8.3', // change the version with the template
   meta: {
     namedImport: true,
   },
