+++ packages/teleport-test/src/standalone.ts
@@ -48,34 +48,34 @@ const run = async () => {
       projectType: ProjectType.REACT,
     })
     console.info(ProjectType.REACT, '-', result.payload)
-    result = await packProject(reactProjectUIDL, {
-      ...packerOptions,
-      projectType: ProjectType.NEXT,
-    })
-    console.info(ProjectType.NEXT, '-', result.payload)
-    result = await packProject(projectUIDL, { ...packerOptions, projectType: ProjectType.NUXT })
-    console.info(ProjectType.NUXT, '-', result.payload)
-    result = await packProject(reactProjectUIDL, { ...packerOptions, projectType: ProjectType.VUE })
-    console.info(ProjectType.VUE, '-', result.payload)
-    result = await packProject(projectUIDL, { ...packerOptions, projectType: ProjectType.STENCIL })
-    console.info(ProjectType.STENCIL, '-', result.payload)
-    result = await packProject(reactProjectUIDL, {
+    // result = await packProject(reactProjectUIDL, {
+    //   ...packerOptions,
+    //   projectType: ProjectType.NEXT,
+    // })
+    // console.info(ProjectType.NEXT, '-', result.payload)
+    // result = await packProject(projectUIDL, { ...packerOptions, projectType: ProjectType.NUXT })
+    // console.info(ProjectType.NUXT, '-', result.payload)
+    // result = await packProject(reactProjectUIDL, { ...packerOptions, projectType: ProjectType.VUE })
+    // console.info(ProjectType.VUE, '-', result.payload)
+    // result = await packProject(projectUIDL, { ...packerOptions, projectType: ProjectType.STENCIL })
+    // console.info(ProjectType.STENCIL, '-', result.payload)
+    result = await packProject(projectUIDL, {
       ...packerOptions,
       projectType: ProjectType.PREACT,
     })
     console.info(ProjectType.PREACT, '-', result.payload)
-    result = await packProject(projectUIDL, {
-      ...packerOptions,
-      projectType: ProjectType.ANGULAR,
-    })
-    console.info(ProjectType.ANGULAR, '-', result.payload)
-    result = await packProject(projectUIDL, { ...packerOptions, projectType: ProjectType.GRIDSOME })
-    console.info(ProjectType.GRIDSOME, '-', result.payload)
-    result = await packProject(reactProjectUIDL, {
-      ...packerOptions,
-      projectType: ProjectType.GATSBY,
-    })
-    console.info(ProjectType.GATSBY, '-', result.payload)
+    // result = await packProject(projectUIDL, {
+    //   ...packerOptions,
+    //   projectType: ProjectType.ANGULAR,
+    // })
+    // console.info(ProjectType.ANGULAR, '-', result.payload)
+    // result = await packProject(projectUIDL, { ...packerOptions, projectType: ProjectType.GRIDSOME })
+    // console.info(ProjectType.GRIDSOME, '-', result.payload)
+    // result = await packProject(reactProjectUIDL, {
+    //   ...packerOptions,
+    //   projectType: ProjectType.GATSBY,
+    // })
+    // console.info(ProjectType.GATSBY, '-', result.payload)
   } catch (e) {
     console.info(e)
   }
